+++ src/de/mknoll/thesis/datastructures/tagcloud/SetDifferenceTagComparatorStrategy.java
@@ -5,7 +5,6 @@ import java.util.Iterator;
 import java.util.List;
 
 import org.mcavallo.opencloud.Tag;
-import org.mcavallo.opencloud.Tag.NameComparatorAsc;
 
 
 
@@ -32,21 +31,21 @@ public class SetDifferenceTagComparatorStrategy implements TagComparatorStrategy
 	 * 
 	 * @Override
 	 */
-	public Float compare(List<Tag> tags1, List<Tag> tags2) {
+	public Double compare(List<Tag> tags1, List<Tag> tags2) {
 		// Make tag lists sorted by ascending tag name
 		Collections.sort(tags1, new Tag.NameComparatorAsc());
 		Collections.sort(tags2, new Tag.NameComparatorAsc());
 		
 		// Assertion: tags are sorted in alphabetical order
 		
-		Float difference = 0F;
+		Double difference = 0D;
 		Iterator<Tag> it1 = tags1.iterator();
 		Iterator<Tag> it2 = tags2.iterator();
 		Tag t1;
 		Tag t2;
 		if (it1.hasNext()) t1 = it1.next(); else t1 = null; 
 		if (it2.hasNext()) t2 = it2.next(); else t2 = null;
-		 while (t1 != null || t2 != null) {
+		while (t1 != null || t2 != null) {
 			if (t1 == null) {
 				difference += 1;
 				if (it2.hasNext()) t2 = it2.next(); else t2 = null;
